+++ ui/src/shared/components/TableGraph.js
@@ -130,16 +130,16 @@ class TableGraph extends Component {
   }
 
   handleHover = (columnIndex, rowIndex) => () => {
-    const {onSetHoverTime, tableOptions: {verticalTimeAxis}} = this.props
+    const {handleSetHoverTime, tableOptions: {verticalTimeAxis}} = this.props
     const {sortedTimeVals} = this.state
     if (verticalTimeAxis && rowIndex === 0) {
       return
     }
-    if (onSetHoverTime) {
+    if (handleSetHoverTime) {
       const hoverTime = verticalTimeAxis
         ? sortedTimeVals[rowIndex]
         : sortedTimeVals[columnIndex]
-      onSetHoverTime(hoverTime.toString())
+      handleSetHoverTime(hoverTime.toString())
     }
     this.setState({
       hoveredColumnIndex: columnIndex,
@@ -148,8 +148,8 @@ class TableGraph extends Component {
   }
 
   handleMouseLeave = () => {
-    if (this.props.onSetHoverTime) {
-      this.props.onSetHoverTime(NULL_HOVER_TIME)
+    if (this.props.handleSetHoverTime) {
+      this.props.handleSetHoverTime(NULL_HOVER_TIME)
       this.setState({
         hoveredColumnIndex: NULL_ARRAY_INDEX,
         hoveredRowIndex: NULL_ARRAY_INDEX,
@@ -222,7 +222,7 @@ class TableGraph extends Component {
       sortField,
       sortDirection,
     } = this.state
-    const {tableOptions, colors} = this.props
+    const {tableOptions, colors} = parent.props
 
     const {
       timeFormat = TIME_FORMAT_DEFAULT,
@@ -233,16 +233,19 @@ class TableGraph extends Component {
 
     const cellData = processedData[rowIndex][columnIndex]
 
-    const timeField = fieldNames.find(
+    const timeFieldIndex = fieldNames.findIndex(
       field => field.internalName === TIME_FIELD_DEFAULT.internalName
     )
-    const visibleTime = _.get(timeField, 'visible', true)
+
+    const visibleTime = _.get(fieldNames, [timeFieldIndex, 'visible'], true)
 
     const isFixedRow = rowIndex === 0 && columnIndex > 0
     const isFixedColumn = fixFirstColumn && rowIndex > 0 && columnIndex === 0
     const isTimeData =
       visibleTime &&
-      (verticalTimeAxis ? rowIndex > 0 && columnIndex === 0 : isFixedRow)
+      (verticalTimeAxis
+        ? rowIndex !== 0 && columnIndex === timeFieldIndex
+        : rowIndex === timeFieldIndex && columnIndex !== 0)
     const isFieldName = verticalTimeAxis ? rowIndex === 0 : columnIndex === 0
     const isFixedCorner = rowIndex === 0 && columnIndex === 0
     const dataIsNumerical = _.isNumber(cellData)
@@ -255,7 +258,7 @@ class TableGraph extends Component {
 
     let cellStyle = style
 
-    if (!isFixedRow && !isFixedColumn && !isFixedCorner) {
+    if (!isFixedRow && !isFixedColumn && !isFixedCorner && !isTimeData) {
       const {bgColor, textColor} = generateThresholdsListHexs({
         colors,
         lastValue: cellData,
@@ -300,7 +303,7 @@ class TableGraph extends Component {
         style={cellStyle}
         className={cellClass}
         onClick={isFieldName ? this.handleClickFieldName(cellData) : null}
-        onMouseOver={this.handleHover(columnIndex, rowIndex)}
+        onMouseOver={_.throttle(this.handleHover(columnIndex, rowIndex), 100)}
         title={cellContents}
       >
         {cellContents}
@@ -406,7 +409,7 @@ TableGraph.propTypes = {
     fixFirstColumn: bool,
   }),
   hoverTime: string,
-  onSetHoverTime: func,
+  handleSetHoverTime: func,
   colors: colorsStringSchema,
   queryASTs: arrayOf(shape()),
 }